refactor project structure and naming
