+++ app/javascript/components/Workbench/QueryEditor/QueryEditor.tsx
@@ -3,7 +3,7 @@ import { Button, Segment, Checkbox } from 'semantic-ui-react';
 import { observer } from 'mobx-react';
 import _ from 'lodash';
 import SqlEditor from './SqlEditor';
-import SqlResults from '../SqlResults';
+import ResultIndex from '../ResultIndex';
 import { default as QueryModel } from '../../../models/QueryEditor';
 import { REST } from '../../../declarations/REST';
 import Slider from '../../../shared/Slider';
@@ -83,7 +83,7 @@ export default class QueryEditor extends React.Component<Props> {
             Query
           </Button>
         </div>
-        <SqlResults query={this.props.query} />
+        <ResultIndex query={this.props.query} />
       </Fragment>
     );
   }
